requirements.txt
pyperclip==1.8.2
keyboard==0.13.5
google-generativeai==0.8.2
pynput==1.8.1
